Fixed the form, now submits POST requests successfuly.

+++ app/booking_views.py
@@ -74,17 +74,21 @@ def dashboard():
 
 @booking_views.route('/tickets', methods=['GET', 'POST'])
 def tickets():
+
     form = TicketForm()
-    if form.validate_on_submit():
-        booking_chosen = Booking.query.get(form.booking_id.data)
-        print(form.refund.data, form.activate.data)
-        if form.refund.data:
-            print("=================== refund was pushed")
-            refund(booking_chosen.payment_intent)
-            db.session.delete(booking_chosen)
-            db.session.commit()
-        elif form.activate.data:
-            redirect(url_for('activate'))
-            print("=================== refund was pushed")
+    if request.method == "POST":
+        if form.validate_on_submit():
+            booking_chosen = Booking.query.get(form.booking_id.data)
+            print(form.refund.data, form.activate.data)
+            if form.refund.data:
+                print("=================== refund was pushed")
+                refund(booking_chosen.payment_intent)
+                db.session.delete(booking_chosen)
+                db.session.commit()
+            elif form.activate.data:
+                print("=================== activate was pushed")
+                return redirect(url_for('activate'))
+        else:
+            flash_errors(form)
 
     return render_template('tickets.html', form=form, page_name='tickets', bookings=current_user.bookings)
